restructure steps
