+++ questverse-widgets/apps/QuestVerse/widget/components/quest/create/step1.jsx
@@ -38,89 +38,6 @@ return (
         </span>
         DAO Info & KYC
       </h2>
-      <Widget
-        src="nearui.near/widget/Input.ExperimentalText"
-        props={{
-          label: "DAO Name",
-          placeholder: "Lorem Ipsum",
-          size: "md",
-          onChange: (v) => {
-            onValueChange("name", v);
-            // generate address
-            onValueChange(
-              "address",
-              `${v
-                .toLowerCase()
-                .replace(/\s/g, "-")
-                .replace(/[^a-zA-Z0-9-]/g, "")}.sputnik-dao.near`,
-            );
-          },
-          inputProps: {
-            name: "name",
-            defaultValue: state.answers.name,
-          },
-          error: errors["name"],
-        }}
-      />
-      <Widget
-        src="nearui.near/widget/Input.ExperimentalText"
-        props={{
-          label: (
-            <>
-              DAO Address{" "}
-              <span className="text-black-50 fw-light small">
-                (auto-filled)
-              </span>
-            </>
-          ),
-          placeholder: "sample-dao-name.sputnik-dao.near",
-          value:
-            state.answers.address === "" ? undefined : state.answers.address,
-          size: "md",
-          disabled: true,
-          onChange: (v) => onValueChange("address", v),
-          inputProps: {
-            name: "address",
-            defaultValue: state.answers.address,
-          },
-          error: errors["address"],
-        }}
-      />
-      <Widget
-        src="nearui.near/widget/Input.ExperimentalText"
-        props={{
-          label: (
-            <>
-              Soul Bound Token Issuer{" "}
-              <span className="text-black-50 fw-light small">(optional)</span>
-            </>
-          ),
-          placeholder: "The address of the token issuer",
-          size: "md",
-          onChange: (v) => onValueChange("soulBoundTokenIssuer", v),
-          error: errors["soulBoundTokenIssuer"],
-          inputProps: {
-            name: "soulBoundTokenIssuer",
-            defaultValue: state.answers.soulBoundTokenIssuer,
-          },
-        }}
-      />
-      <Widget
-        src="nearui.near/widget/Input.ExperimentalText"
-        props={{
-          label: "Purpose",
-          placeholder: "Please add the purpose of your DAO",
-          size: "md",
-          textarea: true,
-          inputProps: {
-            rows: 5,
-            name: "purpose",
-            defaultValue: state.answers.purpose,
-          },
-          onChange: (v) => onValueChange("purpose", v),
-          error: errors["purpose"],
-        }}
-      />
       <h3 className="h5 fw-bold">
         KYC <span className="text-black-50 fw-light small">(optional)</span>
       </h3>
